rename from "l4c" to "do"
l4c was a bad name. the "4" came from forth because the language was
initially meant to be a one-letter-per-instruction forth-like language
like IBNIZ. the "c" came from C because the language was since going to
be GLSL/C like. and now it looks like go!

I'm pretty sold on "do", pronounced as the "do" in solfege "do re me.."
so it varies with language, but rhymes with "go" in english :) I'm also
happy about its petty problems; people are likely to mispronounce it
"doo", it's probably even harder to google than "go", and it's a
reserved keyword in many languages including bash :)